backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic[email]
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
